Moves archive to internal service
